+++ database/couchdb_client.py
