+++ database/crud/todos.py
@@ -1,7 +1,7 @@
 import datetime
 import logging
 import uuid
-from modules.couchdb_client import couchdb_client
+from database.couchdb_client import couchdb_client
 
 logger = logging.getLogger("uvicorn.error.todos")
