+++ modules/twitch_api/twitch_client.py
@@ -10,7 +10,7 @@ from modules.twitch_api import (
     TwitchRewards
 )
 
-logger = logging.getLogger("uvicorn.error.twitch_api")
+logger = logging.getLogger("uvicorn.error.twitch_api_client")
 
 REAL_SCOPES = [
     AuthScope.ANALYTICS_READ_EXTENSION,
